Added glad library
Adding the glad library took a bit of tinkering on my part because I didn't know how to do it with only CMake at first but now it's done. I even managed to make CMake generate a file that clangd can read to update its parsing so that the newly added glad does not show up in red in Neovim.